# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_mongoose.ObjectIdToString_mongoose.FlattenMaps_mongoose.Require_id_DocType-and-TVirtuals___.Exclude_keyofmongoose.ObjectIdToString_mongoose.FlattenMaps_mongoose.Require_id_DocType-and-TVirtuals___.__v__": {
				"properties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_mongoose.ObjectIdToString_mongoose.FlattenMaps_mongoose.Require_id_DocType-and-TVirtuals___.__v_": {
				"$ref": "#/components/schemas/Pick_mongoose.ObjectIdToString_mongoose.FlattenMaps_mongoose.Require_id_DocType-and-TVirtuals___.Exclude_keyofmongoose.ObjectIdToString_mongoose.FlattenMaps_mongoose.Require_id_DocType-and-TVirtuals___.__v__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"mongoose.Document_unknown.__.IUser.__.___": {
				"$ref": "#/components/schemas/Omit_mongoose.ObjectIdToString_mongoose.FlattenMaps_mongoose.Require_id_DocType-and-TVirtuals___.__v_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"IUser": {
				"$ref": "#/components/schemas/Omit_mongoose.ObjectIdToString_mongoose.FlattenMaps_mongoose.Require_id_DocType-and-TVirtuals___.__v_"
			},
			"Required___id-mongoose.Types.ObjectId__": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "df-2.0-live-chat-with-websocket",
		"version": "1.0.0",
		"description": "A one page live chat application",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/mongoose.Document_unknown.__.IUser.__.___"
										},
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"$ref": "#/components/schemas/Required___id-mongoose.Types.ObjectId__"
										},
										{
											"properties": {
												"__v": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"__v"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/mongoose.Document_unknown.__.IUser.__.___"
										},
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"$ref": "#/components/schemas/Required___id-mongoose.Types.ObjectId__"
										},
										{
											"properties": {
												"__v": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"__v"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/upgrade/{userId}": {
			"post": {
				"operationId": "UpgradeToAgent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/mongoose.Document_unknown.__.IUser.__.___"
										},
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"$ref": "#/components/schemas/Required___id-mongoose.Types.ObjectId__"
										},
										{
											"properties": {
												"__v": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"__v"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/role/{role}": {
			"get": {
				"operationId": "FetchUsersByRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"$ref": "#/components/schemas/mongoose.Document_unknown.__.IUser.__.___"
											},
											{
												"$ref": "#/components/schemas/IUser"
											},
											{
												"$ref": "#/components/schemas/Required___id-mongoose.Types.ObjectId__"
											},
											{
												"properties": {
													"__v": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"__v"
												],
												"type": "object"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "role",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}